backend/requirements.txt
Django==5.1.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
drf-yasg==1.21.7
djangorestframework-simplejwt==5.3.0
stripe==7.8.0
python-dotenv==1.0.0
firebase-admin==6.4.0
google-cloud-firestore==2.13.1
